BitStream construction improvements
- added built-in samples for DD and HD bitstreams.
- added more building helpers to TrackBuffer.
- changed BitstreamTrackBuffer to use BitBuffer for storage.
- removed unused FluxTrackBuffer.h from KF and SCP.
- moved more forced unsigned types to be int.

+++ include/CRC16.h
@@ -16,8 +16,8 @@ public:
 	operator uint16_t () const;
 
 	void init(uint16_t crc = INIT_CRC);
-	uint16_t add (uint8_t byte);
-	uint16_t add (uint8_t byte, size_t len);
+	uint16_t add (int byte);
+	uint16_t add (int byte, size_t len);
 	uint16_t add (const void *buf, size_t len);
 
 private:
